usb-host: address release blockers
