+++ embassy-stm32/src/usb/usb_host.rs
@@ -623,30 +623,6 @@ impl<'d, I: Instance, T: channel::Type, D: channel::Direction> UsbChannel<T, D> 
         Ok(())
     }
 
-    fn retarget_channel(
-        &mut self,
-        addr: u8,
-        endpoint: &embassy_usb_driver::EndpointInfo,
-        _pre: bool,
-    ) -> Result<(), embassy_usb_driver::host::HostError> {
-        trace!(
-            "retarget_channel: addr: {:?} ep_type: {:?} index: {}",
-            addr, endpoint.ep_type, self.index
-        );
-        let eptype = endpoint.ep_type;
-        let index = self.index;
-
-        // configure channel register
-        let epr_reg = I::regs().epr(index);
-        let mut epr = invariant(epr_reg.read());
-        epr.set_devaddr(addr);
-        epr.set_ep_type(convert_type(eptype));
-        epr.set_ea(index as _);
-        epr_reg.write_value(epr);
-
-        Ok(())
-    }
-
     async fn request_in(&mut self, buf: &mut [u8]) -> Result<usize, ChannelError>
     where
         D: channel::IsIn,
@@ -661,7 +637,10 @@ impl<'d, I: Instance, T: channel::Type, D: channel::Direction> UsbChannel<T, D> 
         self.write(buf, ensure_transaction_end).await
     }
 
-    fn set_timeout(&mut self, _: TimeoutConfig) {
+    fn set_timeout(&mut self, _: TimeoutConfig)
+    where
+        T: channel::IsControl,
+    {
         //TODO: Implement.
     }
 }
@@ -684,7 +663,7 @@ impl<'d, I: Instance> UsbHostDriver for UsbHost<'d, I> {
         &self,
         addr: u8,
         endpoint: &embassy_usb_driver::EndpointInfo,
-        pre: bool,
+        _pre: bool,
     ) -> Result<Self::Channel<T, D>, embassy_usb_driver::host::HostError> {
         let new_index = if T::ep_type() == EndpointType::Control {
             // Only a single control channel is available
@@ -733,16 +712,21 @@ impl<'d, I: Instance> UsbHostDriver for UsbHost<'d, I> {
             None
         };
 
-        let mut channel = Channel::<I, D, T>::new(
+        // configure channel register
+        let epr_reg = I::regs().epr(new_index as usize);
+        let mut epr = invariant(epr_reg.read());
+        epr.set_devaddr(addr);
+        epr.set_ep_type(convert_type(endpoint.ep_type));
+        epr.set_ea(new_index as _);
+        epr_reg.write_value(epr);
+
+        Ok(Channel::new(
             new_index as usize,
             buffer_in,
             buffer_out,
             endpoint.max_packet_size,
             endpoint.max_packet_size,
-        );
-
-        channel.retarget_channel(addr, endpoint, pre)?;
-        Ok(channel)
+        ))
     }
 
     async fn bus_reset(&self) {
